refactor: Refactored and removed unused function
Refactored the following:
1. Removed the files used for handling the cart and wishlist (replaced with API calls)
2. Removed the unused cases from the reducer
