+++ src/helper/itemInCartWishlist.js
@@ -2,7 +2,6 @@ const itemInCartWishlist = (productId, alreadyPresentProducts) => {
   const isProductPresent = alreadyPresentProducts.some(
     (item) => item._id === productId
   );
-  console.log(isProductPresent);
   return isProductPresent;
 };
