+++ src/reducer/cartWishlistReducer.js
@@ -1,13 +1,3 @@
-// import {
-//   addToCart,
-//   addToWishlist,
-//   removeFromCart,
-//   removeFromWishlist,
-//   moveToCart,
-//   incrementQuantity,
-//   decrementQuantity,
-//   moveToWishlist,
-// } from "../helper";
 import {
   CART_LOADING,
   ADD_TO_CART,
@@ -48,7 +38,6 @@ const cartWishlistReducer = (state, action) => {
         cart: [...action.payload],
       };
     case CHANGE_CART_QUANTITY:
-      console.log(action.payload);
       return {
         ...state,
         cartLoading: false,
@@ -89,34 +78,6 @@ const cartWishlistReducer = (state, action) => {
         ...state,
         wishlistError: "There was some error in processing your request",
       };
-    // case "MOVE_TO_CART":
-    //   const [updatedCart, updatedWishlist] = moveToCart(
-    //     action.payload,
-    //     state.cart,
-    //     state.wishlist
-    //   );
-    //   return {
-    //     ...state,
-    //     cart: [...updatedCart],
-    //     wishlist: [...updatedWishlist],
-    //   };
-    // case "INCREASE_QTY":
-    //   return {
-    //     ...state,
-    //     cart: [...incrementQuantity(action.payload, state.cart)],
-    //   };
-    // case "DECREASE_QTY":
-    //   return {
-    //     ...state,
-    //     cart: [...decrementQuantity(action.payload, state.cart)],
-    //   };
-    // case "MOVE_TO_WISHLIST":
-    //   const [newCart, newWishlist] = moveToWishlist(
-    //     action.payload,
-    //     state.wishlist,
-    //     state.cart
-    //   );
-    // return { ...state, wishlist: [...newWishlist], cart: [...newCart] };
     case "TOTAL_PRICE":
       return { ...state, totalPrice: action.payload };
     default:
